# backend/openapi.json
{
  "components": {
    "schemas": {
      "AnswerReceiptView": {
        "properties": {
          "accepted": {
            "description": "ผลถูก/ผิดเปิดเผยตอนเฉลย ผ่าน state",
            "type": "boolean"
          },
          "optionId": {
            "format": "uuid",
            "type": "string"
          },
          "questionIndex": {
            "type": "number"
          }
        },
        "required": [
          "questionIndex",
          "optionId",
          "accepted"
        ],
        "type": "object"
      },
...
      "BankItemView": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "difficulty": {
            "enum": [
              "EASY",
              "MEDIUM",
              "HARD"
            ],
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "imageUrl": {
            "nullable": true,
            "type": "string"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/OptionView"
            },
            "type": "array"
          },
          "ownerCoreUserId": {
            "type": "string"
          },
          "points": {
            "type": "number"
          },
          "prompt": {
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "timeLimit": {
            "type": "number"
          },
          "type": {
            "enum": [
              "MULTIPLE_CHOICE",
              "TRUE_FALSE"
            ],
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "usedCount": {
            "description": "จำนวนครั้งที่ถูกคัดลอกไปใช้ในแบบทดสอบ",
            "type": "number"
          }
        },
        "required": [
          "id",
          "ownerCoreUserId",
          "type",
          "prompt",
          "timeLimit",
          "points",
          "tags",
          "difficulty",
          "usedCount",
          "options",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
      "CreateBankItemDto": {
        "properties": {
          "difficulty": {
            "default": "MEDIUM",
            "enum": [
              "EASY",
              "MEDIUM",
              "HARD"
            ],
            "type": "string"
          },
          "imageUrl": {
            "description": "ลิงก์รูปภาพ https:// เท่านั้น",
            "nullable": true,
            "type": "string"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/OptionInputDto"
            },
            "maxItems": 4,
            "minItems": 2,
            "type": "array"
          },
          "points": {
            "description": "0 = ไม่คิดคะแนน",
            "maximum": 5000,
            "minimum": 0,
            "type": "number"
          },
          "prompt": {
            "description": "แบบร่างเว้นว่างได้ แต่เผยแพร่ต้องมี",
            "maxLength": 300,
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "type": "array"
          },
          "timeLimit": {
            "description": "วินาที",
            "maximum": 300,
            "minimum": 5,
            "type": "number"
          },
          "type": {
            "enum": [
              "MULTIPLE_CHOICE",
              "TRUE_FALSE"
            ],
            "type": "string"
          }
        },
        "required": [
          "type",
          "prompt",
          "timeLimit",
          "points",
          "options"
        ],
        "type": "object"
      },
...
      "GameJoinView": {
        "properties": {
          "avatarIndex": {
            "type": "number"
          },
          "gameSessionId": {
            "format": "uuid",
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "playerId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "gameSessionId",
          "playerId",
          "nickname",
          "avatarIndex"
        ],
        "type": "object"
      },
      "GameLobbyView": {
        "properties": {
          "gamePin": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "joined": {
            "description": "ผู้เรียกเข้าห้องนี้แล้ว",
            "type": "boolean"
          },
          "playerCount": {
            "type": "number"
          },
          "quizTitle": {
            "type": "string"
          },
          "status": {
            "enum": [
              "LOBBY",
              "ACTIVE",
              "FINISHED",
              "CANCELLED"
            ],
            "type": "string"
          },
          "totalQuestions": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "gamePin",
          "quizTitle",
          "status",
          "playerCount",
          "totalQuestions",
          "joined"
        ],
        "type": "object"
      },
      "GameMeView": {
        "properties": {
          "answer": {
            "allOf": [
              {
                "$ref": "#/components/schemas/GameMyAnswerView"
              }
            ],
            "nullable": true,
            "type": "object"
          },
          "avatarIndex": {
            "type": "number"
          },
          "nickname": {
            "type": "string"
          },
          "playerId": {
            "format": "uuid",
            "type": "string"
          },
          "rank": {
            "type": "number"
          },
          "score": {
            "type": "number"
          },
          "streak": {
            "type": "number"
          }
        },
        "required": [
          "playerId",
          "nickname",
          "avatarIndex",
          "score",
          "streak",
          "rank"
        ],
        "type": "object"
      },
...
      "GamePlayerView": {
        "properties": {
          "avatarIndex": {
            "type": "number"
          },
          "hasAnswered": {
            "description": "ตอบข้อปัจจุบันแล้ว",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "rank": {
            "type": "number"
          },
          "score": {
            "type": "number"
          },
          "streak": {
            "description": "ตอบถูกติดกันตอนนี้",
            "type": "number"
          }
        },
        "required": [
          "id",
          "nickname",
          "avatarIndex",
          "score",
          "streak",
          "rank",
          "hasAnswered"
        ],
        "type": "object"
      },
...
      "GameStateView": {
        "properties": {
          "answeredCount": {
            "type": "number"
          },
          "countdownEndsAt": {
            "description": "นับถอยหลังเต็มจอ: ปิดรับคำตอบ (QUESTION) หรือ ข้อถัดไป/ประกาศผล (LEADERBOARD)",
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "currentQuestionIndex": {
            "type": "number"
          },
          "distribution": {
            "additionalProperties": {
              "type": "number"
            },
            "description": "จำนวนคนที่เลือกแต่ละตัวเลือก { optionId: count } — ผู้เล่นเห็นหลังเฉลย",
            "nullable": true,
            "type": "object"
          },
          "finishedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "gamePin": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "me": {
            "allOf": [
              {
                "$ref": "#/components/schemas/GameMeView"
              }
            ],
            "nullable": true,
            "type": "object"
          },
          "phase": {
            "enum": [
              "LOBBY",
              "QUESTION",
              "RESULT",
              "LEADERBOARD",
              "PODIUM"
            ],
            "type": "string"
          },
          "phaseEndsAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "playerCount": {
            "type": "number"
          },
          "players": {
            "description": "เรียงตามอันดับ",
            "items": {
              "$ref": "#/components/schemas/GamePlayerView"
            },
            "type": "array"
          },
          "question": {
            "allOf": [
              {
                "$ref": "#/components/schemas/GameQuestionView"
              }
            ],
            "nullable": true,
            "type": "object"
          },
          "questionStartedAt": {
            "description": "เวลาเริ่มนับเวลาตอบ (อาจอยู่ในอนาคตระหว่างนับ 3-2-1 ก่อนข้อแรก)",
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "quizId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "quizTitle": {
            "type": "string"
          },
          "results": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PodiumResultsView"
              }
            ],
            "description": "สถิติและรางวัลพิเศษ — มีเฉพาะเมื่อ phase = PODIUM",
            "nullable": true,
            "type": "object"
          },
          "serverTime": {
            "description": "เวลาของ server ใช้ชดเชยนาฬิกาเครื่องผู้ใช้",
            "format": "date-time",
            "type": "string"
          },
          "settings": {
            "$ref": "#/components/schemas/GameSettingsView"
          },
          "startedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "status": {
            "enum": [
              "LOBBY",
              "ACTIVE",
              "FINISHED",
              "CANCELLED"
            ],
            "type": "string"
          },
          "totalQuestions": {
            "type": "number"
          },
          "viewer": {
            "enum": [
              "HOST",
              "PLAYER"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "gamePin",
          "status",
          "phase",
          "viewer",
          "quizTitle",
          "currentQuestionIndex",
          "totalQuestions",
          "serverTime",
          "settings",
          "answeredCount",
          "playerCount",
          "players",
          "createdAt"
        ],
        "type": "object"
      },
      "JoinGameDto": {
        "properties": {
          "avatarIndex": {
            "default": 0,
            "maximum": 47,
            "minimum": 0,
            "type": "number"
          },
          "nickname": {
            "maxLength": 20,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "nickname"
        ],
        "type": "object"
      },
...
      "OptionInputDto": {
        "properties": {
          "id": {
            "description": "id เดิม (ถ้ามี) — ไม่ส่ง = ตัวเลือกใหม่",
            "format": "uuid",
            "type": "string"
          },
          "isCorrect": {
            "type": "boolean"
          },
          "text": {
            "maxLength": 120,
            "type": "string"
          }
        },
        "required": [
          "text",
          "isCorrect"
        ],
        "type": "object"
      },
      "OptionView": {
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "isCorrect": {
            "type": "boolean"
          },
          "text": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "text",
          "isCorrect"
        ],
        "type": "object"
      },
...
      "PodiumResultsView": {
        "properties": {
          "awards": {
            "items": {
              "$ref": "#/components/schemas/PodiumAwardView"
            },
            "type": "array"
          },
          "players": {
            "items": {
              "$ref": "#/components/schemas/PodiumPlayerStatsView"
            },
            "type": "array"
          }
        },
        "required": [
          "players",
          "awards"
        ],
        "type": "object"
      },
...
      "SubmitAnswerDto": {
        "properties": {
          "optionId": {
            "description": "ตัวเลือกของคำถามข้อปัจจุบัน",
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "optionId"
        ],
        "type": "object"
      },
      "UpdateBankItemDto": {
        "properties": {
          "difficulty": {
            "default": "MEDIUM",
            "enum": [
              "EASY",
              "MEDIUM",
              "HARD"
            ],
            "type": "string"
          },
          "imageUrl": {
            "description": "ลิงก์รูปภาพ https:// เท่านั้น",
            "nullable": true,
            "type": "string"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/OptionInputDto"
            },
            "maxItems": 4,
            "minItems": 2,
            "type": "array"
          },
          "points": {
            "description": "0 = ไม่คิดคะแนน",
            "maximum": 5000,
            "minimum": 0,
            "type": "number"
          },
          "prompt": {
            "description": "แบบร่างเว้นว่างได้ แต่เผยแพร่ต้องมี",
            "maxLength": 300,
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "type": "array"
          },
          "timeLimit": {
            "description": "วินาที",
            "maximum": 300,
            "minimum": 5,
            "type": "number"
          },
          "type": {
            "enum": [
              "MULTIPLE_CHOICE",
              "TRUE_FALSE"
            ],
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "csmju_quiz_access_token": {
        "in": "cookie",
        "name": "csmju_quiz_access_token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "API ของระบบย่อย CSMJU Quiz — ทุก response ห่อด้วย { success, data[, meta] } หรือ { success: false, error }",
    "title": "CSMJU Quiz API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/bank-items": {
      "get": {
        "operationId": "BankItemsController_list",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "maxLength": 30,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "difficulty",
            "required": false,
            "schema": {
              "enum": [
                "EASY",
                "MEDIUM",
                "HARD"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BankItemView"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "csmju_quiz_access_token": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "List my question bank items",
        "tags": [
          "bank-items"
        ]
      },
      "post": {
        "operationId": "BankItemsController_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBankItemDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankItemView"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "csmju_quiz_access_token": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Add a complete question to my bank",
        "tags": [
          "bank-items"
        ]
      }
    },
    "/api/v1/bank-items/{id}": {
      "delete": {
        "operationId": "BankItemsController_remove",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "csmju_quiz_access_token": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Delete from bank (quiz copies are kept)",
        "tags": [
          "bank-items"
        ]
      },
      "get": {
        "operationId": "BankItemsController_findOne",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankItemView"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "csmju_quiz_access_token": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "bank-items"
        ]
      },
      "patch": {
        "operationId": "BankItemsController_update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBankItemDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankItemView"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "csmju_quiz_access_token": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "bank-items"
        ]
      }
    },
...
    "/api/v1/guest-games": {
      "get": {
        "operationId": "GuestGamesController_lobby",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "ค้นห้องที่เปิดรอผู้เล่นด้วยรหัสเกม (ผู้เล่นใช้ค่านี้) · ไม่ส่ง = ห้องของฉัน (host)",
            "in": "query",
            "name": "gamePin",
            "required": false,
            "schema": {
              "pattern": "^\\d{6}$",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "LOBBY",
                "ACTIVE",
                "FINISHED",
                "CANCELLED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GameLobbyView"
                  },
                  "type": "array"
                }
              }
            },
            "description": "GameLobbyView[] (0–1 ห้อง)"
          }
        },
        "security": [
          {
            "quiz_room_pass": []
          }
        ],
        "summary": "Guest: find the room waiting for players with ?gamePin=",
        "tags": [
          "guest-games"
        ]
      }
    },
    "/api/v1/guest-games/{id}": {
      "get": {
        "operationId": "GuestGamesController_findOne",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateView"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "quiz_room_pass": []
          }
        ],
        "summary": "Guest: current game state (room pass required)",
        "tags": [
          "guest-games"
        ]
      }
    },
    "/api/v1/guest-games/{id}/answers": {
      "post": {
        "operationId": "GuestGamesController_answer",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitAnswerDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerReceiptView"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "quiz_room_pass": []
          }
        ],
        "summary": "Guest: submit an answer to the current question",
        "tags": [
          "guest-games"
        ]
      }
    },
    "/api/v1/guest-games/{id}/events": {
      "get": {
        "operationId": "GuestGamesController_events",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "quiz_room_pass": []
          }
        ],
        "summary": "Guest: live game state stream (text/event-stream)",
        "tags": [
          "guest-games"
        ]
      }
    },
    "/api/v1/guest-games/{id}/players": {
      "post": {
        "operationId": "GuestGamesController_join",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinGameDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameJoinView"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "quiz_room_pass": []
          }
        ],
        "summary": "Guest: join a room in the lobby — sets the room pass cookie (HttpOnly)",
        "tags": [
          "guest-games"
        ]
      }
    },
    "/api/v1/guest-games/{id}/players/{playerId}": {
      "delete": {
        "operationId": "GuestGamesController_leave",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "playerId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "quiz_room_pass": []
          }
        ],
        "summary": "Guest: leave the lobby (own player only)",
        "tags": [
          "guest-games"
        ]
      }
    },
...
  },
  "servers": [],
...
}